src/examples/no-graphygie-neo4j/pyproject.toml
[project]
name = "neo4j-graphrag-example"
version = "0.1.0"
authors = [
    { name = "c2fc2f", email = "contact@c2fc2f.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain-ollama>=1.0.0",
    "neo4j>=5.28.2",
    "neo4j-graphrag>=1.10.1",
    "ollama>=0.6.0",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "regex>=2025.10.23",
    "tqdm>=4.67.1",
]

[tool.hatch.build.targets.wheel]
packages = [
  "./main.py"
]

[build-system]
requires = [
  "hatchling"
]
build-backend = "hatchling.build"
